+++ structures/imagingtable.cpp
@@ -100,7 +100,7 @@ void ImagingTable::AssignGridDataFromPolarization(
 }
 
 std::unique_ptr<DeconvolutionTable> ImagingTable::CreateDeconvolutionTable(
-    size_t n_deconvolution_channels, CachedImageSet& psf_images,
+    int n_deconvolution_channels, CachedImageSet& psf_images,
     CachedImageSet& model_images, CachedImageSet& residual_images) const {
   // In a DeconvolutionTable the output channel indices range from
   // 0 to (#channels - 1). In an ImagingTable that forms an indepent group,
@@ -108,8 +108,8 @@ std::unique_ptr<DeconvolutionTable> ImagingTable::CreateDeconvolutionTable(
 
   // Assume that the first entry has the lowest index and that the last entry
   // has the highest output channel index.
-  const size_t channel_index_offset = _entries.front()->outputChannelIndex;
-  const size_t n_original_channels =
+  const int channel_index_offset = _entries.front()->outputChannelIndex;
+  const int n_original_channels =
       _entries.back()->outputChannelIndex + 1 - channel_index_offset;
 
   auto table = std::make_unique<DeconvolutionTable>(
